+++ docs/source/releases/changelog.rst
@@ -7,6 +7,15 @@ Tracpro's version is incremented upon each merge to master according to our
 We recommend reviewing the release notes and code diffs before upgrading
 between versions.
 
+v1.4.0 (released 2016-03-28)
+----------------------------
+
+Code diff: https://github.com/rapidpro/tracpro/compare/v1.3.3...v1.4.0
+
+* Migrations to move RapidPro uuid unique constraint to unique_together with
+  another model field (`org` for Contact, Region, and Group models; `pollrun`
+  for Response)
+
 v1.3.3 (released 2016-03-28)
 ----------------------------